+++ Tests/Unit/Controller/BackEnd/EventControllerTest.php
@@ -12,15 +12,21 @@ use OliverKlee\Seminars\Domain\Repository\Event\EventRepository;
 use OliverKlee\Seminars\Service\EventStatisticsCalculator;
 use OliverKlee\Seminars\Tests\Unit\Controller\RedirectMockTrait;
 use PHPUnit\Framework\MockObject\MockObject;
-use TYPO3\CMS\Backend\Template\ModuleTemplate;
 use TYPO3\CMS\Backend\Template\ModuleTemplateFactory;
+use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
+use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
 use TYPO3\CMS\Core\Http\HtmlResponse;
 use TYPO3\CMS\Core\Http\RedirectResponse;
 use TYPO3\CMS\Core\Http\ServerRequest;
+use TYPO3\CMS\Core\Imaging\IconFactory;
 use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Localization\LanguageService;
+use TYPO3\CMS\Core\Messaging\FlashMessageService;
+use TYPO3\CMS\Core\Package\PackageManager;
+use TYPO3\CMS\Core\Page\PageRenderer;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
+use TYPO3\CMS\Fluid\View\StandaloneView;
 use TYPO3\CMS\Fluid\View\TemplateView;
 use TYPO3\TestingFramework\Core\AccessibleObjectInterface;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
@@ -32,6 +38,8 @@ final class EventControllerTest extends UnitTestCase
 {
     use RedirectMockTrait;
 
+    protected bool $resetSingletonInstances = true;
+
     /**
      * @var EventController&MockObject&AccessibleObjectInterface
      */
@@ -41,16 +49,6 @@ final class EventControllerTest extends UnitTestCase
      * @var TemplateView&MockObject
      */
     private TemplateView $viewMock;
-
-    /**
-     * @var ModuleTemplateFactory&MockObject
-     */
-    private ModuleTemplateFactory $moduleTemplateFactoryMock;
-
-    /**
-     * @var ModuleTemplate&MockObject
-     */
-    private ModuleTemplate $moduleTemplateMock;
 
     /**
      * @var EventRepository&MockObject
@@ -81,12 +79,10 @@ final class EventControllerTest extends UnitTestCase
     {
         parent::setUp();
 
-        $this->moduleTemplateFactoryMock = $this->createMock(ModuleTemplateFactory::class);
+        $moduleTemplateFactory = $this->createModuleTemplateFactory();
         $this->eventRepositoryMock = $this->createMock(EventRepository::class);
         $this->permissionsMock = $this->createMock(Permissions::class);
         $this->eventStatisticsCalculatorMock = $this->createMock(EventStatisticsCalculator::class);
-        $this->languageServiceMock = $this->createMock(LanguageService::class);
-        $GLOBALS['LANG'] = $this->languageServiceMock;
 
         $methodsToMock = ['addFlashMessage', 'htmlResponse', 'redirect', 'redirectToUri'];
         /** @var EventController&AccessibleObjectInterface&MockObject $subject */
@@ -94,7 +90,7 @@ final class EventControllerTest extends UnitTestCase
             EventController::class,
             $methodsToMock,
             [
-                $this->moduleTemplateFactoryMock,
+                $moduleTemplateFactory,
                 $this->eventRepositoryMock,
                 $this->permissionsMock,
                 $this->eventStatisticsCalculatorMock,
@@ -104,8 +100,6 @@ final class EventControllerTest extends UnitTestCase
 
         $request = $this->createStub(ServerRequest::class);
         $this->subject->_set('request', $request);
-        $this->moduleTemplateMock = $this->createMock(ModuleTemplate::class);
-        $this->moduleTemplateFactoryMock->method('create')->with($request)->willReturn($this->moduleTemplateMock);
 
         $responseStub = $this->createStub(HtmlResponse::class);
         $this->subject->method('htmlResponse')->willReturn($responseStub);
@@ -120,10 +114,37 @@ final class EventControllerTest extends UnitTestCase
 
     protected function tearDown(): void
     {
-        unset($_GET['id'], $GLOBALS['LANG']);
+        unset($_GET['id'], $GLOBALS['LANG'], $GLOBALS['BE_USER']);
         GeneralUtility::purgeInstances();
 
         parent::tearDown();
+    }
+
+    /**
+     * Note: This is a real mockfest. We need to convert the BE controller tests to functional tests first.
+     */
+    private function createModuleTemplateFactory(): ModuleTemplateFactory
+    {
+        GeneralUtility::addInstance(StandaloneView::class, $this->createStub(StandaloneView::class));
+        $packageManagerMock = $this->createMock(PackageManager::class);
+        $packageManagerMock->method('getActivePackages')->willReturn([]);
+        GeneralUtility::setSingletonInstance(PackageManager::class, $packageManagerMock);
+        $this->languageServiceMock = $this->createMock(LanguageService::class);
+        $this->languageServiceMock->lang = 'default';
+        $GLOBALS['LANG'] = $this->languageServiceMock;
+        $extensionConfigurationMock = $this->createMock(ExtensionConfiguration::class);
+        $extensionConfigurationMock->method('get')->with('backend', 'backendFavicon')->willReturn('icon.svg');
+        GeneralUtility::addInstance(ExtensionConfiguration::class, $extensionConfigurationMock);
+        $GLOBALS['BE_USER'] = $this->createStub(BackendUserAuthentication::class);
+
+        $pageRenderMock = $this->createMock(PageRenderer::class);
+        $pageRenderMock->method('getLanguage')->willReturn('default');
+
+        return new ModuleTemplateFactory(
+            $pageRenderMock,
+            $this->createStub(IconFactory::class),
+            $this->createStub(FlashMessageService::class)
+        );
     }
 
     /**
@@ -239,9 +260,6 @@ final class EventControllerTest extends UnitTestCase
      */
     public function searchActionReturnsHtmlResponse(): void
     {
-        $this->moduleTemplateMock->expects(self::once())->method('setContent')
-            ->with('rendered view');
-
         $result = $this->subject->searchAction(1, '');
 
         self::assertInstanceOf(HtmlResponse::class, $result);