+++ Tests/Unit/Controller/BackEnd/ModuleControllerTest.php
@@ -12,11 +12,19 @@ use OliverKlee\Seminars\Domain\Repository\Registration\RegistrationRepository;
 use OliverKlee\Seminars\Service\EventStatisticsCalculator;
 use OliverKlee\Seminars\Tests\Unit\Controller\RedirectMockTrait;
 use PHPUnit\Framework\MockObject\MockObject;
-use TYPO3\CMS\Backend\Template\ModuleTemplate;
 use TYPO3\CMS\Backend\Template\ModuleTemplateFactory;
+use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
+use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
 use TYPO3\CMS\Core\Http\HtmlResponse;
 use TYPO3\CMS\Core\Http\ServerRequest;
+use TYPO3\CMS\Core\Imaging\IconFactory;
+use TYPO3\CMS\Core\Localization\LanguageService;
+use TYPO3\CMS\Core\Messaging\FlashMessageService;
+use TYPO3\CMS\Core\Package\PackageManager;
+use TYPO3\CMS\Core\Page\PageRenderer;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
+use TYPO3\CMS\Fluid\View\StandaloneView;
 use TYPO3\CMS\Fluid\View\TemplateView;
 use TYPO3\TestingFramework\Core\AccessibleObjectInterface;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
@@ -30,6 +38,8 @@ use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 final class ModuleControllerTest extends UnitTestCase
 {
     use RedirectMockTrait;
+
+    protected bool $resetSingletonInstances = true;
 
     /**
      * @var ModuleController&MockObject&AccessibleObjectInterface
@@ -52,16 +62,6 @@ final class ModuleControllerTest extends UnitTestCase
     private EventRepository $eventRepositoryMock;
 
     /**
-     * @var ModuleTemplateFactory&MockObject
-     */
-    private ModuleTemplateFactory $moduleTemplateFactoryMock;
-
-    /**
-     * @var ModuleTemplate&MockObject
-     */
-    private ModuleTemplate $moduleTemplateMock;
-
-    /**
      * @var RegistrationRepository&MockObject
      */
     private RegistrationRepository $registrationRepositoryMock;
@@ -75,7 +75,7 @@ final class ModuleControllerTest extends UnitTestCase
     {
         parent::setUp();
 
-        $this->moduleTemplateFactoryMock = $this->createMock(ModuleTemplateFactory::class);
+        $moduleTemplateFactory = $this->createModuleTemplateFactory();
         $this->eventRepositoryMock = $this->createMock(EventRepository::class);
         $this->registrationRepositoryMock = $this->createMock(RegistrationRepository::class);
 
@@ -84,14 +84,12 @@ final class ModuleControllerTest extends UnitTestCase
         $subject = $this->getAccessibleMock(
             ModuleController::class,
             $methodsToMock,
-            [$this->moduleTemplateFactoryMock, $this->eventRepositoryMock, $this->registrationRepositoryMock]
+            [$moduleTemplateFactory, $this->eventRepositoryMock, $this->registrationRepositoryMock]
         );
         $this->subject = $subject;
 
         $request = $this->createStub(ServerRequest::class);
         $this->subject->_set('request', $request);
-        $this->moduleTemplateMock = $this->createMock(ModuleTemplate::class);
-        $this->moduleTemplateFactoryMock->method('create')->with($request)->willReturn($this->moduleTemplateMock);
 
         $responseStub = $this->createStub(HtmlResponse::class);
         $this->subject->method('htmlResponse')->willReturn($responseStub);
@@ -107,9 +105,37 @@ final class ModuleControllerTest extends UnitTestCase
 
     protected function tearDown(): void
     {
-        unset($_GET['id'], $_GET['pid']);
+        unset($_GET['id'], $_GET['pid'], $GLOBALS['LANG'], $GLOBALS['BE_USER']);
+        GeneralUtility::purgeInstances();
 
         parent::tearDown();
+    }
+
+    /**
+     * Note: This is a real mockfest. We need to convert the BE controller tests to functional tests first.
+     */
+    private function createModuleTemplateFactory(): ModuleTemplateFactory
+    {
+        GeneralUtility::addInstance(StandaloneView::class, $this->createStub(StandaloneView::class));
+        $packageManagerMock = $this->createMock(PackageManager::class);
+        $packageManagerMock->method('getActivePackages')->willReturn([]);
+        GeneralUtility::setSingletonInstance(PackageManager::class, $packageManagerMock);
+        $languageServiceStub = $this->createStub(LanguageService::class);
+        $languageServiceStub->lang = 'default';
+        $GLOBALS['LANG'] = $languageServiceStub;
+        $extensionConfigurationMock = $this->createMock(ExtensionConfiguration::class);
+        $extensionConfigurationMock->method('get')->with('backend', 'backendFavicon')->willReturn('icon.svg');
+        GeneralUtility::addInstance(ExtensionConfiguration::class, $extensionConfigurationMock);
+        $GLOBALS['BE_USER'] = $this->createStub(BackendUserAuthentication::class);
+
+        $pageRenderMock = $this->createMock(PageRenderer::class);
+        $pageRenderMock->method('getLanguage')->willReturn('default');
+
+        return new ModuleTemplateFactory(
+            $pageRenderMock,
+            $this->createStub(IconFactory::class),
+            $this->createStub(FlashMessageService::class)
+        );
     }
 
     /**
@@ -155,9 +181,6 @@ final class ModuleControllerTest extends UnitTestCase
      */
     public function overviewActionReturnsHtmlResponse(): void
     {
-        $this->moduleTemplateMock->expects(self::once())->method('setContent')
-            ->with('rendered view');
-
         $result = $this->subject->overviewAction();
 
         self::assertInstanceOf(HtmlResponse::class, $result);